+++ data/src/main/java/com/moneymong/moneymong/data/datasource/AgencyRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package com.moneymong.moneymong.network.datasource
+package com.moneymong.moneymong.data.datasource
 
 import com.moneymong.moneymong.network.request.agency.AgencyRegisterRequest
