+++ data/src/main/java/com/moneymong/moneymong/data/di/DataSourceModule.kt
@@ -1,7 +1,7 @@
-package com.moneymong.moneymong.network.di
+package com.moneymong.moneymong.data.di
 
-import com.moneymong.moneymong.network.datasource.AgencyRemoteDataSource
-import com.moneymong.moneymong.network.datasource.mock.AgencyRemoteDataSourceMock
+import com.moneymong.moneymong.data.datasource.AgencyRemoteDataSource
+import com.moneymong.moneymong.data.datasource.mock.AgencyRemoteDataSourceMock
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -9,7 +9,7 @@ import dagger.hilt.components.SingletonComponent
 
 @Module
 @InstallIn(SingletonComponent::class)
-interface SourceModule {
+interface DataSourceModule {
 
     @Binds
     fun bindAgencyRemoteDataSource(
